Separate .h and .cpp
