+++ core/gdxsv/gdxsv_backend_rollback.cpp
@@ -21,7 +21,7 @@
 
 namespace {
 u8 DummyGameParam[] = {0x00, 0x00, 0x01, 0x00, 0x03, 0x00, 0x02, 0x00, 0x05, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x83,
-						  0x76, 0x83, 0x8c, 0x83, 0x43, 0x83, 0x84, 0x81, 0x5b, 0x82, 0x50, 0x00, 0x00, 0x00, 0x00, 0x07};
+					   0x76, 0x83, 0x8c, 0x83, 0x43, 0x83, 0x84, 0x81, 0x5b, 0x82, 0x50, 0x00, 0x00, 0x00, 0x00, 0x07};
 u8 DummyRuleData[] = {0x03, 0x02, 0x03, 0x00, 0x00, 0x01, 0x58, 0x02, 0x58, 0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xff, 0xff, 0xff,
 					  0x3f, 0xff, 0xff, 0xff, 0x3f, 0x00, 0x00, 0xff, 0x01, 0xff, 0xff, 0xff, 0x3f, 0xff, 0xff, 0xff, 0x3f, 0x00};
 
@@ -242,7 +242,6 @@ void GdxsvBackendRollback::OnMainUiLoop() {
 	}
 
 	if (is_local_test_ && state_ == State::End) {
-		SaveReplay();
 		dc_exit();
 	}
 }
@@ -287,11 +286,11 @@ bool GdxsvBackendRollback::StartLocalTest(const char* param) {
 	Prepare(matching, 20010 + me);
 	state_ = State::StartLocalTest;
 	is_local_test_ = true;
-	maxlag_ = 0;
-	maxrebattle_ = 1;
+	gdxsv.maxlag = 0;
+	gdxsv.maxrebattle = 1;
 
 	if (getenv("MAXREBATTLE")) {
-		maxrebattle_ = atoi(getenv("MAXREBATTLE"));
+		gdxsv.maxrebattle = atoi(getenv("MAXREBATTLE"));
 	}
 
 	NOTICE_LOG(COMMON, "RollbackNet StartLocalTest %d", me);
@@ -325,7 +324,7 @@ void GdxsvBackendRollback::Open() {
 	NOTICE_LOG(COMMON, "GdxsvBackendRollback.Open");
 	recv_buf_.assign({0x0e, 0x61, 0x00, 0x22, 0x10, 0x31, 0x66, 0x77, 0x88, 0x99, 0xaa, 0xbb, 0xcc, 0xdd});
 	state_ = State::McsSessionExchange;
-	maxlag_ = 0;
+	gdxsv.maxlag = 0;
 	ApplyPatch(true);
 }
 
@@ -383,7 +382,7 @@ u32 GdxsvBackendRollback::OnSockRead(u32 addr, u32 size) {
 		}
 
 		auto inputState = mapleInputState;
-		auto memExInputAddr = symbols_.at("print_buf");	 // TODO
+		auto memExInputAddr = gdxsv.symbols.at("print_buf");  // TODO
 
 		int msg_len = gdxsv_ReadMem8(InetBuf);
 		if (0 < msg_len) {
